+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## CHANGELOG
 
 1. some refactoring.
-2. added subreddit names debounced search !
+2. improved subreddit names debounced search.